// helpers/types.ts
export enum ProtocolErrors {
  //common errors
  CALLER_NOT_POOL_ADMIN = "100", // 'The caller must be the pool admin'
  CALLER_NOT_ADDRESS_PROVIDER = "101",
  INVALID_FROM_BALANCE_AFTER_TRANSFER = "102",
  INVALID_TO_BALANCE_AFTER_TRANSFER = "103",
  CALLER_NOT_ONBEHALFOF_OR_IN_WHITELIST = "104",

  //math library erros
  MATH_MULTIPLICATION_OVERFLOW = "200",
  MATH_ADDITION_OVERFLOW = "201",
  MATH_DIVISION_BY_ZERO = "202",

  //validation & check errors
  VL_INVALID_AMOUNT = "301", // 'Amount must be greater than 0'
  VL_NO_ACTIVE_RESERVE = "302", // 'Action requires an active reserve'
  VL_RESERVE_FROZEN = "303", // 'Action cannot be performed because the reserve is frozen'
  VL_NOT_ENOUGH_AVAILABLE_USER_BALANCE = "304", // 'User cannot withdraw more than the available balance'
  VL_BORROWING_NOT_ENABLED = "305", // 'Borrowing is not enabled'
  VL_COLLATERAL_BALANCE_IS_0 = "306", // 'The collateral balance is 0'
  VL_HEALTH_FACTOR_LOWER_THAN_LIQUIDATION_THRESHOLD = "307", // 'Health factor is lesser than the liquidation threshold'
  VL_COLLATERAL_CANNOT_COVER_NEW_BORROW = "308", // 'There is not enough collateral to cover a new borrow'
  VL_NO_DEBT_OF_SELECTED_TYPE = "309", // 'for repayment of stable debt, the user needs to have stable debt, otherwise, he needs to have variable debt'
  VL_NO_ACTIVE_NFT = "310",
  VL_NFT_FROZEN = "311",
  VL_SPECIFIED_CURRENCY_NOT_BORROWED_BY_USER = "312", // 'User did not borrow the specified currency'
  VL_INVALID_HEALTH_FACTOR = "313",
  VL_INVALID_ONBEHALFOF_ADDRESS = "314",
  VL_INVALID_TARGET_ADDRESS = "315",
  VL_INVALID_RESERVE_ADDRESS = "316",
  VL_SPECIFIED_LOAN_NOT_BORROWED_BY_USER = "317",
  VL_SPECIFIED_RESERVE_NOT_BORROWED_BY_USER = "318",

  //lend pool errors
  LP_CALLER_NOT_LEND_POOL_CONFIGURATOR = "400", // 'The caller of the function is not the lending pool configurator'
  LP_IS_PAUSED = "401", // 'Pool is paused'
  LP_NO_MORE_RESERVES_ALLOWED = "402",
  LP_NOT_CONTRACT = "403",
  LP_BORROW_NOT_EXCEED_LIQUIDATION_THRESHOLD = "404",
  LP_BORROW_IS_EXCEED_LIQUIDATION_PRICE = "405",
  LP_NO_MORE_NFTS_ALLOWED = "406",
  LP_INVALIED_USER_NFT_AMOUNT = "407",
  LP_INCONSISTENT_PARAMS = "408",
  LP_NFT_IS_NOT_USED_AS_COLLATERAL = "409",
  LP_CALLER_MUST_BE_AN_CTOKEN = "410",
  LP_INVALIED_NFT_AMOUNT = "411",
  LP_NFT_HAS_USED_AS_COLLATERAL = "412",
  LP_DELEGATE_CALL_FAILED = "413",
  LP_AMOUNT_LESS_THAN_EXTRA_DEBT = "414",
  LP_AMOUNT_LESS_THAN_REDEEM_THRESHOLD = "415",
  LP_AMOUNT_GREATER_THAN_MAX_REPAY = "416",

  //lend pool loan errors
  LPL_INVALID_LOAN_STATE = "480",
  LPL_INVALID_LOAN_AMOUNT = "481",
  LPL_INVALID_TAKEN_AMOUNT = "482",
  LPL_AMOUNT_OVERFLOW = "483",
  LPL_BUYITNOW_PRICE_LESS_THAN_LIQUIDATION_PRICE = "484",
  LPL_BUYITNOW_PRICE_LESS_THAN_HIGHEST_PRICE = "485",
  LPL_BUYITNOW_REDEEM_DURATION_HAS_END = "486",
  LPL_BUYITNOW_USER_NOT_SAME = "487",
  LPL_BUYITNOW_REPAY_AMOUNT_NOT_ENOUGH = "488",
  LPL_BUYITNOW_DURATION_HAS_END = "489",
  LPL_BUYITNOW_DURATION_NOT_END = "490",
  LPL_BUYITNOW_PRICE_LESS_THAN_BORROW = "491",
  LPL_INVALID_BUYITNOW_BUYER_ADDRESS = "492",
  LPL_AMOUNT_LESS_THAN_BUYITNOW_FINE = "493",
  LPL_INVALID_BUYITNOW_FINE = "494",
  LPL_BUYITNOW_DURATION_NOT_START = "495",

  //common token errors
  CT_CALLER_MUST_BE_LEND_POOL = "500", // 'The caller of this function must be a lending pool'
  CT_INVALID_MINT_AMOUNT = "501", //invalid amount to mint
  CT_INVALID_BURN_AMOUNT = "502", //invalid amount to burn
  CT_BORROW_ALLOWANCE_NOT_ENOUGH = "503",

  //reserve logic errors
  RL_RESERVE_ALREADY_INITIALIZED = "601", // 'Reserve has already been initialized'
  RL_LIQUIDITY_INDEX_OVERFLOW = "602", //  Liquidity index overflows uint128
  RL_VARIABLE_BORROW_INDEX_OVERFLOW = "603", //  Variable borrow index overflows uint128
  RL_LIQUIDITY_RATE_OVERFLOW = "604", //  Liquidity rate overflows uint128
  RL_VARIABLE_BORROW_RATE_OVERFLOW = "605", //  Variable borrow rate overflows uint128

  //configure errors
  LPC_RESERVE_LIQUIDITY_NOT_0 = "700", // 'The liquidity of the reserve needs to be 0'
  LPC_INVALID_CONFIGURATION = "701", // 'Invalid risk parameters for the reserve'
  LPC_CALLER_NOT_EMERGENCY_ADMIN = "702", // 'The caller must be the emergency admin'
  LPC_INVALIED_CNFT_ADDRESS = "703",
  LPC_INVALIED_LOAN_ADDRESS = "704",
  LPC_NFT_LIQUIDITY_NOT_0 = "705",

  //reserve config errors
  RC_INVALID_LTV = "730",
  RC_INVALID_LIQ_THRESHOLD = "731",
  RC_INVALID_LIQ_BONUS = "732",
  RC_INVALID_DECIMALS = "733",
  RC_INVALID_RESERVE_FACTOR = "734",
  RC_INVALID_REDEEM_DURATION = "735",
  RC_INVALID_BUYITNOW_DURATION = "736",
  RC_INVALID_REDEEM_FINE = "737",

  //address provider erros
  LPAPR_PROVIDER_NOT_REGISTERED = "760", // 'Provider is not registered'
  LPAPR_INVALID_ADDRESSES_PROVIDER_ID = "761",

  // Misc errors
  INVALID_OWNER_REVERT_MSG = "Ownable: caller is not the owner",
  TRANSFER_AMOUNT_EXCEEDS_BALANCE = "ERC20: transfer amount exceeds balance",
  SAFEERC20_LOWLEVEL_CALL = "SafeERC20: low-level call failed",
}